Authenticate and Create a case
These are all the objects needed to create a case.
They are all unit tested and a high level feature spec is required when
we come to invoking them in the background job.

The gateway still has some knowledge of a "Complaint" which will be
removed in a future PR.

In rails 6 this required moving to the app directory due to changes in how autoloading works
The only other option was to disable Zeitwerk using `config.autoloader = :classic`