Record a successful submission
Delayed::Job doesn't have this metric out of the box since jobs are not soft
deleted. We want to keep track of this metric to ensure the form is working well.

A future PR will expose this data on a /metrics endpoint which can be viewed
manually or piped into a dashboard like Grafana.

Don't throw an error when failing to record a successful submission,
this will cause a job to retry and re-send a submission to Optics