Use the newer ECR repos
We needed to change the name of the ECR repositories that this app uses.
Unfortunately this takes the app away from the consitency of the other
Form Builder applications, all of which start with `fb-`.

The reason for this is because changing the deployment.yaml to adjust
the pod name is a more dangerous thing to do than change which ECR repo
is pushed and pulled from.

The new fb-deploy scripts rely heavily on naming consistency for each
app between the ECR repo as well as the image used in the deployment.