Add secrets to deployments
This brings the adapter into line with the other apps on the platform.

Previously the adapter secrets were held as an actual kubernetes secret
config file which the pipeline applied.

Now the secrets are mapped to a `secrets.yaml` file in this repo and the
fb-deploy scripts will apply them during deployment.